styler app/requirements.txt
flask
torch
torchvision
matplotlib
torch-summary
gunicorn
pillow
